TODO: rewrite CharRep to not be existent
The CharRep struct is annoying because it needs to be updated manually.
Instead, I should rewrite it so that data is a Vec<u8> and so that I can
display the bytes either as the two hex digits or as ASCII depending on
the necessary context. Hope to get that done by tonight